Activated onClick on widget and opening detail UI

+++ app/src/main/java/org/checklist/comics/comicschecklist/ui/ActivityAddComic.java
@@ -60,7 +60,7 @@ public class ActivityAddComic extends AppCompatActivity {
 
         // Create the detail fragment and add it to the activity using a fragment transaction.
         Bundle arguments = new Bundle();
-        arguments.putLong(Constants.ARG_COMIC_ID, getIntent().getLongExtra(Constants.ARG_COMIC_ID, -1));
+        arguments.putInt(Constants.ARG_COMIC_ID, getIntent().getIntExtra(Constants.ARG_COMIC_ID, -1));
         FragmentAddComic fragment = new FragmentAddComic();
         fragment.setArguments(arguments);
         getSupportFragmentManager().beginTransaction().add(R.id.comic_detail_container, fragment, "addComicFragment").commit();
